feat:auto send settle pr tx
